+++ ecosystem/typescript/sdk/src/providers/utils.ts
